removed duplicate functions
